Add processing-tests/common and processing-tests/processor-tests
