--- processing-tests/build.gradle.kts
@@ -1,16 +1,0 @@
-plugins {
-    kotlin("kapt")
-}
-
-dependencies {
-    testImplementation(Dependencies.junit)
-    testImplementation(Dependencies.kotlinCompileTesting)
-    testImplementation(kotlin("reflect"))
-    testImplementation(project(":runtime"))
-    testImplementation(project(":processor"))
-    kaptTest(project(":processor"))
-}
-
-kapt {
-    includeCompileClasspath = false
-}
