+++ processing-tests/common/test/java/com/livefront/sealedenum/compilation/kitchensink/JavaInterface2.java
