+++ processing-tests/common/test/java/com/livefront/sealedenum/compilation/kitchensink/JavaInterface3.java
