+++ processing-tests/common/test/java/com/livefront/sealedenum/compilation/kitchensink/JavaSecondBaseClass.java
