+++ processing-tests/common/test/java/com/livefront/sealedenum/compilation/visibility/JavaProtectedInterfaceSubclass.java
