+++ processing-tests/common/test/java/com/livefront/sealedenum/compilation/visibility/subpackage/JavaProtectedInterfaceSubclass.java
