+++ processing-tests/common/test/kotlin/com/livefront/sealedenum/compilation/kitchensink/JavaBaseClasses.kt
