+++ processing-tests/common/test/kotlin/com/livefront/sealedenum/compilation/kitchensink/JavaBaseClassesTests.kt
