+++ processing-tests/common/test/kotlin/com/livefront/sealedenum/compilation/location/NestedClass.kt
