+++ processing-tests/common/test/kotlin/com/livefront/sealedenum/compilation/location/NestedClassTests.kt
