+++ processing-tests/common/test/kotlin/com/livefront/sealedenum/compilation/visibility/ProtectedInterfaceSealedClass.kt
@@ -1,0 +1,164 @@
+package com.livefront.sealedenum.compilation.visibility
+
+import com.livefront.sealedenum.GenSealedEnum
+import org.intellij.lang.annotations.Language
+
+open class ProtectedInterfaceOuterClass {
+
+    protected interface ProtectedInterface
+
+    sealed class ProtectedInterfaceSealedClass :
+        com.livefront.sealedenum.compilation.visibility.JavaProtectedInterfaceSubclass(), ProtectedInterface {
+
+        object FirstObject : ProtectedInterfaceSealedClass()
+
+        object SecondObject : ProtectedInterfaceSealedClass()
+
+        @GenSealedEnum(generateEnum = true)
+        companion object
+    }
+}
+
+@Language("kotlin")
+val protectedInterfaceSealedClassGenerated = """
+package com.livefront.sealedenum.compilation.visibility
+
+import com.livefront.sealedenum.EnumForSealedEnumProvider
+import com.livefront.sealedenum.SealedEnum
+import com.livefront.sealedenum.SealedEnumWithEnumProvider
+import java.lang.Class
+import kotlin.Int
+import kotlin.String
+import kotlin.collections.List
+
+/**
+ * An isomorphic enum for the sealed class
+ * [ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass]
+ */
+public enum class ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum(
+    sealedObject: ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass
+) : JavaProtectedInterfaceBaseClass.ProtectedInterface by sealedObject {
+    ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_FirstObject(com.livefront.sealedenum.compilation.visibility.ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.FirstObject),
+    ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_SecondObject(com.livefront.sealedenum.compilation.visibility.ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.SecondObject),
+}
+
+/**
+ * The isomorphic [ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum] for [this].
+ */
+public val ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.`enum`:
+        ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum
+    get() =
+            ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassSealedEnum.sealedObjectToEnum(this)
+
+/**
+ * The isomorphic [ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass] for [this].
+ */
+public val ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum.sealedObject:
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass
+    get() =
+            ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassSealedEnum.enumToSealedObject(this)
+
+/**
+ * An implementation of [SealedEnum] for the sealed class
+ * [ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass]
+ */
+public object ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassSealedEnum :
+        SealedEnum<ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass>,
+        SealedEnumWithEnumProvider<ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass,
+        ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum>,
+        EnumForSealedEnumProvider<ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass,
+        ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum> {
+    public override val values: List<ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass> =
+            listOf(
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.FirstObject,
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.SecondObject
+    )
+
+
+    public override val enumClass:
+            Class<ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum>
+        get() = ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum::class.java
+
+    public override fun ordinalOf(obj: ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass):
+            Int = when (obj) {
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.FirstObject -> 0
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.SecondObject -> 1
+    }
+
+    public override fun nameOf(obj: ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass):
+            String = when (obj) {
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.FirstObject ->
+                "ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_FirstObject"
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.SecondObject ->
+                "ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_SecondObject"
+    }
+
+    public override fun valueOf(name: String):
+            ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass = when (name) {
+        "ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_FirstObject" ->
+                ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.FirstObject
+        "ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_SecondObject" ->
+                ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.SecondObject
+        else -> throw IllegalArgumentException(""${'"'}No sealed enum constant ${'$'}name""${'"'})
+    }
+
+    public override
+            fun sealedObjectToEnum(obj: ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass):
+            ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum = when (obj) {
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.FirstObject ->
+                ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum.ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_FirstObject
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.SecondObject ->
+                ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum.ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_SecondObject
+    }
+
+    public override
+            fun enumToSealedObject(`enum`: ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum):
+            ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass = when (enum) {
+        ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum.ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_FirstObject ->
+                ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.FirstObject
+        ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassEnum.ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClass_SecondObject ->
+                ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.SecondObject
+    }
+}
+
+/**
+ * The index of [this] in the values list.
+ */
+public val ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.ordinal: Int
+    get() = ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassSealedEnum.ordinalOf(this)
+
+/**
+ * The name of [this] for use with valueOf.
+ */
+public val ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.name: String
+    get() = ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassSealedEnum.nameOf(this)
+
+/**
+ * A list of all [ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass] objects.
+ */
+public val ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.Companion.values:
+        List<ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass>
+    get() = ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassSealedEnum.values
+
+/**
+ * Returns an implementation of [SealedEnum] for the sealed class
+ * [ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass]
+ */
+public val ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.Companion.sealedEnum:
+        ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassSealedEnum
+    get() = ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassSealedEnum
+
+/**
+ * Returns the [ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass] object for the given
+ * [name].
+ *
+ * If the given name doesn't correspond to any
+ * [ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass], an [IllegalArgumentException] will be
+ * thrown.
+ */
+public
+        fun ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass.Companion.valueOf(name: String):
+        ProtectedInterfaceOuterClass.ProtectedInterfaceSealedClass =
+        ProtectedInterfaceOuterClass_ProtectedInterfaceSealedClassSealedEnum.valueOf(name)
+
+""".trimIndent()
