+++ processing-tests/common/test/kotlin/com/livefront/sealedenum/errors/NonObjectErrors.kt
