+++ processing-tests/common/test/kotlin/com/livefront/sealedenum/testing/Result.kt
@@ -1,0 +1,11 @@
+package com.livefront.sealedenum.testing
+
+import com.tschuchort.compiletesting.KotlinCompilation
+
+/**
+ * A wrapper around [KotlinCompilation.Result] that includes [messages] separately.
+ */
+class CompilationResult(
+    val kotlinCompilationResult: KotlinCompilation.Result,
+    val messages: String
+)
