+++ processing-tests/processor-tests/src/test/kotlin/com/livefront/sealedenum/testing/CurrentProcessingType.kt
@@ -1,0 +1,6 @@
+package com.livefront.sealedenum.testing
+
+/**
+ * Returns the current processing type.
+ */
+val ProcessingType.Companion.currentType get(): ProcessingType = ProcessingType.AnnotationProcessing
