+++ settings.gradle.kts
@@ -1,5 +1,5 @@
 include(":runtime")
 include(":processing-common")
 include(":processor")
-include(":processing-tests")
+include(":processing-tests:processor-tests")
 rootProject.name = "sealed-enum"
